Move the handler owners' receiver-free rules out (#680)
Sixth semantic vertical, taking the 18 candidates #678 deferred. Fourteen
moved; four stayed, and why they stayed is the useful part.

Reductions, with every ledger baseline tightened to the validated figure:

- handlers/quest 8,907 to 8,486 production lines (19,731 to 19,232 total)
- handlers/character 20,266 to 20,205 (33,623 to 33,519)
- handlers/loot 13,939 to 13,914 (30,417 to 30,352)
- session 83,160 to 83,158, world-server totals corrected downward too

`WorldSession` associated items 3,698 to 3,684, with nothing added; fields,
registry accesses and bridge accesses unchanged, because no state moved.

## What the owner's vocabulary turned out to be

The seventeen `*_LIKE_CPP_LOCAL` objective-type constants looked like the
rules' private vocabulary, so the first attempt moved them along. They are
not: `objectives.rs` and `persistence.rs` read them too, so the move broke
the owner. They stayed, widened to `pub(crate)`, and the rules import them
by name - which is the honest description of a shared vocabulary.

Four rules could not leave at all:

- `represented_quest_dialog_classification_like_cpp` and
  `read_quest_choice_item_like_cpp` construct owner structs field by field,
  and `creature_addon_create_fields_like_cpp` does the same for a creature
  addon. Moving them would have meant making those structs' fields
  `pub(crate)` - opening encapsulation to relocate code, which is the
  opposite of the point.
- `represented_reward_choice_matches_loaded_type_like_cpp` reads the same
  private fields.

They returned to their owners, each folded into that file's existing
`impl WorldSession` block rather than appended in a new one, so no physical
ceiling moved. A returned rule also has to lose `Self::` at call sites
outside an impl, which the compiler reported as E0433.

3,822 wow-world lib tests pass unchanged, the workspace build is clean, and
architecture check, self-test, physical ratchet, hotspot ratchet and the
syntax-only session ownership check pass. Five of the eight curated owners
have now moved off their baselines.

Closes #680
