+++ crates/wow-world/src/handlers/loot/mod.rs
@@ -141,9 +141,9 @@ const MAX_NR_LOOT_ITEMS_LIKE_CPP: usize = 18;
 const LOOT_ROLL_TIMEOUT_MS_LIKE_CPP: u32 = 60_000;
 #[cfg(test)]
 const ROLL_ALL_TYPE_NO_DISENCHANT_LIKE_CPP: u8 = 0x07;
-const ROLL_ALL_TYPE_MASK_LIKE_CPP: u8 = 0x0F;
-const ROLL_FLAG_TYPE_NEED_LIKE_CPP: u8 = 0x02;
-const ROLL_FLAG_TYPE_DISENCHANT_LIKE_CPP: u8 = 0x08;
+pub(crate) const ROLL_ALL_TYPE_MASK_LIKE_CPP: u8 = 0x0F;
+pub(crate) const ROLL_FLAG_TYPE_NEED_LIKE_CPP: u8 = 0x02;
+pub(crate) const ROLL_FLAG_TYPE_DISENCHANT_LIKE_CPP: u8 = 0x08;
 const LOOT_SLOT_TYPE_ALLOW_LOOT_LIKE_CPP: u8 = 0;
 const LOOT_SLOT_TYPE_ROLL_ONGOING_LIKE_CPP: u8 = 1;
 const LOOT_SLOT_TYPE_LOCKED_LIKE_CPP: u8 = 2;
